Issue:
On EditButton click, editPasswordForm is getting cloned from the the addPasswordForm. All fields in the editPasswordForm
takes values from the addPasswordForm at that point in time. If the addPasswordForm passwordField is updated, the updated
password is reflected in the editPasswordForm passwordField as well for that siteuser.

Fix:
Set passwordField to the password value obtained from db at the time of clone like it is done for siteUsername and website fields.
This ensures that the password displayed comes from db instead of the current value of password in addPasswordForm.

In addition to that, close the addPasswordForm on submit and clear out all fields on addPasswordButton click event.
Updated the add_password function in the test to reflect the changes for the fix.